+++ src/actions/likeVideoAction.js
@@ -3,6 +3,9 @@ import axios from "axios";
 
 export const GetLikes = createAsyncThunk("likes/get", async () => {
   const token = localStorage.getItem("token");
+  if(!token) {
+    throw new Error(alert("Please Login First"))
+  }
   try {
     const { data } = await axios.get("/api/user/likes", {
       headers: {
@@ -19,6 +22,9 @@ export const AddToLikes = createAsyncThunk(
   "likes/addVideo", 
   async ( video ) => {
     const token = localStorage.getItem("token");
+    if(!token) {
+      throw new Error(alert("Please Login First"))
+    }
     try {
       const { data } = await axios.post("/api/user/likes", { video }, {
         headers: {
@@ -36,6 +42,9 @@ export const DeleteFromLikes = createAsyncThunk(
   "likes/deleteVideo", 
   async (id) => {
     const token = localStorage.getItem("token");
+    if(!token) {
+      throw new Error(alert("Please Login First"))
+    }
     try {
       const { data } = await axios.delete(`/api/user/likes/${id}`, {
         headers: {
